requirements.txt
gymnasium==1.2.2
highway-env==1.10.2
openai==2.8.1
rich==14.2.0
python-dotenv==1.2.1
pyyaml==6.0.3